Issue #10981: Moved VsiteStorageDefinitionTest to drupal-test-traits
