+++ profile/modules/vsite/tests/src/ExistingSite/VsiteCurrentFilterTest.php
@@ -4,7 +4,6 @@ namespace Drupal\Tests\vsite\ExistingSite;
 
 use Drupal\group\Entity\Group;
 use Drupal\views\Views;
-use weitzman\DrupalTestTraits\ExistingSiteBase;
 
 /**
  * Test the Current Vsite Views Filter.
@@ -14,14 +13,7 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  * @group kernel
  * @coversDefaultClass \Drupal\vsite\Plugin\views\filter\VsiteCurrentFilter
  */
-class VsiteCurrentFilterTest extends ExistingSiteBase {
-
-  /**
-   * Views used by this test.
-   *
-   * @var array
-   */
-  public static $testViews = ['vsite_test_view'];
+class VsiteCurrentFilterTest extends VsiteExistingSiteTestBase {
 
   /**
    * Group dummy content is being assigned (or not) to.
@@ -36,7 +28,6 @@ class VsiteCurrentFilterTest extends ExistingSiteBase {
    * @var \Drupal\node\Entity\Node
    */
   protected $ungroupedNode;
-
 
   /**
    * A grouped node.
@@ -61,15 +52,12 @@ class VsiteCurrentFilterTest extends ExistingSiteBase {
     // Set the current user so group creation can rely on it.
     $this->container->get('current_user')->setAccount($this->createUser());
 
-    /** @var \Drupal\Core\Entity\EntityTypeManagerInterface $entityTypeManager */
-    $entityTypeManager = $this->container->get('entity_type.manager');
-
     /** @var \Drupal\group\Entity\GroupTypeInterface $group_type */
-    $group_type = $entityTypeManager->getStorage('group_type')->load('personal');
+    $group_type = $this->entityTypeManager->getStorage('group_type')->load('personal');
 
     // Enable the user_as_content plugin on the default group type.
     /** @var \Drupal\group\Entity\Storage\GroupContentTypeStorageInterface $storage */
-    $storage = $entityTypeManager->getStorage('group_content_type');
+    $storage = $this->entityTypeManager->getStorage('group_content_type');
     $plugin = $storage->createFromPlugin($group_type, 'group_node:page');
     $plugin->save();
 
@@ -117,7 +105,7 @@ class VsiteCurrentFilterTest extends ExistingSiteBase {
    *   A list of view results.
    */
   protected function getViewResults() {
-    $view = Views::getView(reset($this::$testViews));
+    $view = Views::getView('vsite_test_view');
     $view->setDisplay('page_1');
 
     if ($view->preview()) {
