+++ profile/modules/vsite/tests/src/ExistingSite/VsiteExistingSiteTestBase.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace Drupal\Tests\vsite\ExistingSite;
+
+use weitzman\DrupalTestTraits\ExistingSiteBase;
+
+/**
+ * Base class for vsite tests.
+ */
+abstract class VsiteExistingSiteTestBase extends ExistingSiteBase {
+
+  /**
+   * The entity type manager service.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
+   */
+  protected $entityTypeManager;
+
+  /**
+   * {@inheritdoc}
+   */
+  public function setUp() {
+    parent::setUp();
+
+    $this->entityTypeManager = $this->container->get('entity_type.manager');
+  }
+
+  /**
+   * Creates a group.
+   *
+   * @param array $values
+   *   (optional) The values used to create the entity.
+   *
+   * @return \Drupal\group\Entity\GroupInterface
+   *   The created group entity.
+   */
+  protected function createGroup(array $values = []) {
+    $group = $this->entityTypeManager->getStorage('group')->create($values + [
+      'type' => 'default',
+      'label' => $this->randomMachineName(),
+    ]);
+    $group->enforceIsNew();
+    $group->save();
+    return $group;
+  }
+
+}
